+++ src/app/Components/MauiProgram.cs
@@ -1,5 +1,7 @@
 ﻿using DIPS.Mobile.UI.API.Builder;
+using DIPS.Mobile.UI.API.Library;
 using Microsoft.Extensions.Logging;
+using Microsoft.Maui.LifecycleEvents;
 
 namespace Components;
 
@@ -11,6 +13,20 @@ public static class MauiProgram
         builder
             .UseMauiApp<App>()
             .UseDIPSUI();
+        
+        //Close things when the app is going to background and back again
+#if __ANDROID__
+        builder.ConfigureLifecycleEvents(events =>
+        {
+            events.AddAndroid(android => android.OnResume(_ => DUI.RemoveViewsLocatedOnTopOfPage()));
+        });
+
+#elif __IOS__
+ builder.ConfigureLifecycleEvents(events =>
+        {
+            events.AddiOS(ios => ios.OnResignActivation(_ => DUI.RemoveViewsLocatedOnTopOfPage()));
+        });
+#endif
 #if DEBUG
         builder.Logging.AddDebug();
 #endif