+++ src/app/Playground/MauiProgram.cs
@@ -1,6 +1,8 @@
 ﻿using DIPS.Mobile.UI;
 using DIPS.Mobile.UI.API.Builder;
+using DIPS.Mobile.UI.API.Library;
 using Microsoft.Extensions.Logging;
+using Microsoft.Maui.LifecycleEvents;
 
 namespace Playground;
 
@@ -17,7 +19,6 @@ public static class MauiProgram
                 fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
                 fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
             });
-
 #if DEBUG
         builder.Logging.AddDebug();
 #endif